+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import Layout from '@components/common/Layout';
+import Layout from '@components/layout/Layout';
 import '@styles/globals.css';
 import { NextPage } from 'next';
 import type { AppProps } from 'next/app';
